move rendering from ValueObject to Renderer
